Provide encoding wrapper using standard C types
The zbase32_encode() function provides an interface without the zstr
types so that they don't need to be exposed. Furthermore, it wraps the
simpler b2a() function that assumes you want to encode all the bits in
data. I don't have a use case for encoding less than a full byte, so
this keeps from having the caller pass in something unnecessary.